moved Signal from canmatrix to own package, with tests. Also fixes test test_signal_encode_invalid_named_value expected value
